Organise files and filder
