+++ src/components/DoctorConsole/DocInterface.css
