+++ src/components/DoctorConsole/DocOfficePatient.js
@@ -7,7 +7,7 @@ const DocOfficePatient = ({ patient }) => {
         <div className='DocOfficePatient'>
             <h3>Your current Patient :</h3>
             <div className='DocOfficePatient__current'>
-                <img className='Patient__Img' src={require(`../${patient.pic}`)} alt='patient avatar' />
+                <img className='Patient__Img' src={require(`../../${patient.pic}`)} alt='patient avatar' />
                 <p>{patient.name}</p>
             </div>
         </div>
