+++ src/components/General/Navbar.js
@@ -5,7 +5,7 @@ import './Navbar.css'
 const Navbar = () => {
     return (
         <nav className='Navbar'>
-            <img src={require('../images/doctolib-logo.png')} alt='doctolib-logo'/>
+            <img src={require('../../images/doctolib-logo.png')} alt='doctolib-logo'/>
             <p>Ann O'Nyme</p>
         </nav>
     )
