+++ src/components/WaitingRoom/Leaving.css
