+++ src/components/WaitingRoom/Leaving.js
