+++ src/components/WaitingRoom/MainModal.css
@@ -108,3 +108,7 @@
 .ffjfJJ {
     height: 350px !important;
 }
+
+.rsc, .fqndpX, .fLjYpL, .gOCww, .ffjfJJ, .jmCkbN {
+    box-sizing: 'content-box' !important;
+}
