+++ src/components/WaitingRoom/MainModal.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
-import Chatbot from './Chatbot'
-import ChatbotWelcomeMessage from './ChatbotWelcomeMessage'
+import Chatbot from '../Chatbot/Chatbot'
+import ChatbotWelcomeMessage from '../Chatbot/ChatbotWelcomeMessage'
 import Counter from './Counter'
 import Leaving from './Leaving'
 import Office from './Office'
@@ -20,15 +20,15 @@ const MainModal = ({ patients, timeAppointement, timeAnim, showDoc, time, nbrBef
   const showModal = () => setLeaving(true)
   const hideModal = () => setLeaving(false)
 
-  const botAvatarUrl = require('../images/chatbot.svg')
-  const userAvatarUrl = require('../images/6.png')
+  const botAvatarUrl = require('../../images/chatbot.svg')
+  const userAvatarUrl = require('../../images/6.png')
 
   return (
     <div className='MainModal'>
       <div className='MainModal__veil' style={veil ?{display:'block'}:{display:'none'}}></div>
       <div className='MainModal__title'>
         <div className='MainModal__title__left'>
-          <img src={require('../images/logo.png')} alt='waiting_room_logo' />
+          <img src={require('../../images/logo.png')} alt='waiting_room_logo' />
           <h2 onClick={showDoc}>Virtual Waiting Room</h2>
         </div>
         <div className='MainModal__title__right'>
